pyproject.toml
[project]
name = "ai-artist"
version = "0.1.0"
description = "Autonomous AI artist with LoRA style training"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AI Artist Project" }
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "diffusers>=0.27.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "peft>=0.7.0",
    "safetensors>=0.4.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "structlog>=24.0.0",
    "tqdm>=4.66.0",
]

[project.scripts]
ai-artist = "ai_artist.main:main"
ai-artist-gallery = "ai_artist.gallery.viewer:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.0.0",
    "pre-commit>=3.6.0",
]
training = [
    "wandb>=0.16.0",
    "xformers>=0.0.24",
]
quality = [
    "imagehash>=4.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "unit: marks unit tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [".history"]
